refactor: split the predictor and the data manager into packages
konfai/predictor.py becomes konfai/predictor: output (the sink and its
streaming), ensemble (ModelComposite), loop (the per-rank loop), workflow
(Predictor, build_predict, predict); the package still re-exports Mean,
Median, Concat and Reduction for published configs. konfai/data/
data_manager.py becomes konfai/data/data_manager: groups, order, samples,
subset, sources. Tests import a name from the module that defines it and
patch the module that binds it.
