+++ konfai/data/data_manager/sources.py
@@ -14,40 +14,33 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""Dataset assembly, subset selection, and dataloader orchestration for KonfAI."""
+
+"""The data sources of each workflow: training, prediction, evaluation, dataset preparation."""
 
 import math
 import os
 import random
-import threading
-import traceback
 import warnings
 from abc import ABC, abstractmethod
-from collections.abc import Iterator, Mapping
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from dataclasses import dataclass, replace
+from collections.abc import Mapping
 from functools import partial
 from pathlib import Path
-from typing import TypeAlias, cast
+from typing import cast
 
 import numpy as np
-import torch
-import tqdm
-from torch.cuda import device_count
-from torch.utils import data
-from torch.utils.data import DataLoader, Sampler
+from torch.utils.data import DataLoader
 
-from konfai import konfai_root, konfai_state
+from konfai import konfai_state
 from konfai.data.augmentation import DataAugmentation, DataAugmentationsList
+from konfai.data.data_manager.groups import Group, GroupMetric, GroupOut, _chains
+from konfai.data.data_manager.order import WindowedCaseSampler, _interleaved_case_entries
+from konfai.data.data_manager.samples import _CACHE_ELEMENT_BYTES, DatasetIter, collate_konfai
+from konfai.data.data_manager.subset import PredictionSubset, Subset, TrainSubset
 from konfai.data.patching import DatasetManager, DatasetPatch
 from konfai.data.transform import (
     Expand,
-    LocalityKind,
     Reduce,
     Save,
-    Transform,
-    TransformInverse,
-    TransformLoader,
     Write,
 )
 from konfai.utils.budget import (
@@ -59,862 +52,9 @@ from konfai.utils.budget import (
 from konfai.utils.clock import startup_clock
 from konfai.utils.config import config
 from konfai.utils.dataset import Attribute, Dataset
-from konfai.utils.errors import ConfigError, DatasetManagerError, TransformerError
-from konfai.utils.runtime import State, get_cpu_info, get_memory, get_memory_info, memory_forecast
-from konfai.utils.utils import SUPPORTED_FORMATS, OverlapSpec, resolve_patch, split_path_spec
-
-# A cached case is a float32 tensor (torch's default dtype, and the default TensorCast's target), so
-# bytes are counted at 4/element from the header shape alone, not the on-disk dtype, and without
-# modelling transforms that shrink or grow the cached tensor.
-_CACHE_ELEMENT_BYTES = 4
-
-
-def _cache_worker_count(cpu_count: int, device_count: int) -> int:
-    """Number of caching threads: CPUs shared across devices, but never below one."""
-    divisor = device_count if device_count > 0 else 1
-    return max(1, cpu_count // divisor)
-
-
-def _check_patch_transform_locality(transform: Transform, group_src: str, group_dest: str) -> None:
-    """Reject a transform whose per-patch result cannot equal its case-level result.
-
-    Only POINTWISE and GLOBAL_STAT are correct on one patch (per-patch GLOBAL_STAT means: derive the
-    statistic from this patch; use ``lazy=True`` case-level to feed it the volume's). The messages in
-    ``reasons`` below say why each other kind is rejected. Probed with an empty ``Attribute`` (config
-    time has no case), so an image-decided kind answers WHOLE_VOLUME: the right answer here.
-    """
-    kind = transform.patch_locality(Attribute()).kind
-    if kind in (LocalityKind.POINTWISE, LocalityKind.GLOBAL_STAT):
-        return
-    name = type(transform).__name__
-    location = f"{konfai_root()}.Dataset.groups_src.{group_src}.groups_dest.{group_dest}"
-    reasons = {
-        LocalityKind.HALO: (
-            f"'{name}' reads a neighbourhood around each voxel, so run per-patch it would see no data"
-            " beyond the patch border and corrupt every patch edge. Running it per-patch is not"
-            " supported yet."
-        ),
-        LocalityKind.ORIENTATION: (
-            f"'{name}' reorients its input: applied to one patch it reorients that patch about its own"
-            " extent, which is not the whole volume reoriented and then cut into patches."
-        ),
-        LocalityKind.CROP: (
-            f"'{name}' crops its input to a box measured on the whole volume: applied to one patch it"
-            " crops that patch about its own extent, and cuts the patch grid predictions are"
-            " reassembled onto down to what is left."
-        ),
-        LocalityKind.REGRID: (
-            f"'{name}' resamples its input onto another grid: applied to one patch it would rescale"
-            " that patch about its own extent, or hand back the whole target extent: neither of"
-            " which is the patch grid predictions are reassembled onto."
-        ),
-        LocalityKind.WHOLE_VOLUME: f"'{name}' needs the whole volume.",
-    }
-    raise ConfigError(
-        f"{location}.patch_transforms: {reasons[kind]}",
-        f"Move '{name}' to {location}.transforms, where it runs once on the whole volume.",
-    )
-
-
-def _check_patch_transform_shape(transform: Transform, group_src: str, group_dest: str) -> None:
-    """Reject a patch_transform that resizes the patch it is handed.
-
-    The patch grid is folded from the CASE-level ``transforms`` only (``DatasetManager``), so a
-    patch_transform that changes the spatial shape hands back a patch the batch cannot collate and the
-    ``Accumulator`` cannot write onto the grid. ``_check_patch_transform_locality`` above takes the
-    transform at its word; this is the structural check, and it is asked of ``transform_shape`` (the
-    contract every transform already owes the patch planner), with distinct extents, so a swap or a
-    resize of any single axis shows up. Only the SPATIAL shape is at stake: patching hands
-    ``transform_shape`` the channel-stripped shape, so a transform that changes only the channel count
-    (``OneHot``) is not caught here, and must not be: the grid it feeds is spatial.
-
-    Runs after the locality check, which is what makes the bare probe attribute safe: by here the
-    transform is POINTWISE or GLOBAL_STAT, and the kinds whose ``transform_shape`` needs real geometry
-    (``Resample`` reads ``Spacing``) have already been rejected.
-    """
-    spatial_shape = [7, 11, 13]
-    shape = list(transform.transform_shape(group_src, "", list(spatial_shape), Attribute()))
-    if shape == spatial_shape:
-        return
-    name = type(transform).__name__
-    location = f"{konfai_root()}.Dataset.groups_src.{group_src}.groups_dest.{group_dest}"
-    raise ConfigError(
-        f"{location}.patch_transforms: '{name}' changes the spatial shape of its input"
-        f" ({spatial_shape} -> {shape}), but a patch must keep the shape the patch grid cut it to.",
-        f"Move '{name}' to {location}.transforms, where the patch grid is folded from its"
-        " transform_shape(); a patch_transform must be spatially shape-preserving.",
-    )
-
-
-def _check_patch_transform_invertible(
-    transform: Transform, case_transforms: list[Transform], group_src: str, group_dest: str
-) -> None:
-    """Reject a per-patch global statistic at prediction, whose inverse cannot be reconstructed.
-
-    A ``GLOBAL_STAT`` transform is allowed in ``patch_transforms`` (see
-    ``_check_patch_transform_locality``): run per patch it standardizes each patch by that patch's OWN
-    statistic, which is what asking for it per-patch means: correct, and the deliberate training use.
-    But the per-patch statistic lives in the per-patch attribute scope and never reaches the case
-    attribute, so at prediction the finalize inverse, which seeds every patch from the CASE attribute
-    and pops the statistic: has nothing to pop. Nor could it: the reassembled volume was normalised
-    patch by patch with different coefficients, so a single case-level inverse cannot un-apply it. Refuse
-    here, at config time, rather than fail deep in the inverse with a bare ``NameError``.
-
-    A case-level ``transforms`` entry that derives the SAME statistic rescues it: run once on the whole
-    volume it caches that statistic on the case attribute (``Standardize(lazy=True)`` caches Mean/Std and
-    applies nothing), which the per-patch inverse then inherits and pops. So the patch transform is only
-    un-invertible when nothing case-level captures its statistic.
-
-    Training-only use stays valid: the check is gated on the prediction state, where the inverse actually
-    runs (``RESUME``/``TRAIN`` never invert patch_transforms, and evaluation drops them entirely).
-    """
-    if konfai_state() != str(State.PREDICTION):
-        return
-    if not (isinstance(transform, TransformInverse) and transform.apply_inverse):
-        return
-    locality = transform.patch_locality(Attribute())
-    if locality.kind is not LocalityKind.GLOBAL_STAT:
-        return
-    if any(
-        (case_locality := case.patch_locality(Attribute())).kind is LocalityKind.GLOBAL_STAT
-        and locality.stat_keys <= case_locality.stat_keys
-        for case in case_transforms
-    ):
-        return
-    name = type(transform).__name__
-    location = f"{konfai_root()}.Dataset.groups_src.{group_src}.groups_dest.{group_dest}"
-    raise ConfigError(
-        f"{location}.patch_transforms: '{name}' derives its statistic from each patch, but prediction"
-        " must invert it and a per-patch statistic cannot be un-applied to the reassembled volume.",
-        f"Capture the volume-global statistic case-level instead: put '{name}(lazy=True)' in"
-        f" {location}.transforms (it traverses the whole volume, caches the statistic and applies"
-        f" nothing), and keep '{name}()' in patch_transforms to consume it.",
-    )
-
-
-class GroupTransform:
-    """Collection of transforms attached to one source-to-destination group path."""
-
-    def __init__(
-        self,
-        transforms: dict[str, TransformLoader] | None = {
-            "default|Normalize|Standardize|Unsqueeze|TensorCast|ResampleIsotropic|ResampleResize": TransformLoader()
-        },
-        patch_transforms: dict[str, TransformLoader] | None = {
-            "default|Normalize|Standardize|Unsqueeze|TensorCast|ResampleIsotropic|ResampleResize": TransformLoader()
-        },
-        is_input: bool = True,
-    ) -> None:
-        self._transforms = transforms
-        self._patch_transforms = patch_transforms
-        self.transforms: list[Transform] = []
-        self.patch_transforms: list[Transform] = []
-        self.is_input = is_input
-        self._prepared = False
-
-    def prepare(self, group_src: str, group_dest: str) -> None:
-        # Binds ONCE. A workflow that inspects its chains before handing over to Data.prepare() calls
-        # this first, and Data.prepare() calls it again for every group, so without the guard every
-        # stage is constructed twice, and anything the workflow attached to the first set is silently
-        # thrown away with it. A stage's __init__ is user code and may not be run twice for free.
-        if self._prepared:
-            return
-        self._prepared = True
-        self.transforms = []
-        self.patch_transforms = []
-        if self._transforms is not None:
-            for classpath, transform_loader in self._transforms.items():
-                transform = transform_loader.get_transform(
-                    classpath,
-                    konfai_args=f"{konfai_root()}.Dataset.groups_src.{group_src}.groups_dest.{group_dest}.transforms",
-                    # Past an Expand marker the chain is the copies' draws: a name both packages
-                    # have (Flip, Mask, Permute) is the draw there, the transform before it.
-                    prefer_augmentation=any(isinstance(stage, Expand) for stage in self.transforms),
-                )
-                self.transforms.append(transform)
-        if self._patch_transforms is not None:
-            for classpath, transform_loader in self._patch_transforms.items():
-                transform = transform_loader.get_transform(
-                    classpath,
-                    konfai_args=f"{konfai_root()}.Dataset.groups_src.{group_src}"
-                    f".groups_dest.{group_dest}.patch_transforms",
-                )
-                _check_patch_transform_locality(transform, group_src, group_dest)
-                _check_patch_transform_shape(transform, group_src, group_dest)
-                _check_patch_transform_invertible(transform, self.transforms, group_src, group_dest)
-                self.patch_transforms.append(transform)
-
-    def set_datasets(self, datasets: list[Dataset]) -> None:
-        for transform in self.transforms:
-            transform.set_datasets(datasets)
-        for transform in self.patch_transforms:
-            transform.set_datasets(datasets)
-
-    def to(self, device: int):
-        for transform in self.transforms:
-            transform.to(device)
-        for transform in self.patch_transforms:
-            transform.to(device)
-
-    def __str__(self) -> str:
-        params = {"transforms": self.transforms, "patch_transforms": self.patch_transforms}
-        return str(params)
-
-    def __repr__(self) -> str:
-        return str(self)
-
-
-class GroupTransformMetric(GroupTransform):
-    """Metric-specific group transform that omits patch-time transforms."""
-
-    def __init__(
-        self,
-        transforms: dict[str, TransformLoader] = {
-            "default|Normalize|Standardize|Unsqueeze|TensorCast|ResampleIsotropic|ResampleResize": TransformLoader()
-        },
-    ):
-        super().__init__(transforms, {})
-
-
-class GroupTransformOut(GroupTransform):
-    """Transform-workflow group: a plain chain, no patch-time transforms, no ``is_input``.
-
-    Every group of a dataset-preparation workflow is an input, so the flag is not a question to ask; and the
-    patch grid is an execution detail the planner owns, so a per-patch transform would make the
-    OUTPUT a function of the declared budget."""
-
-    def __init__(
-        self,
-        transforms: dict[str, TransformLoader] = {
-            "default|Normalize|Standardize|TensorCast|ResampleIsotropic|Write": TransformLoader()
-        },
-    ):
-        super().__init__(transforms, {})
-
-
-class Group(dict[str, GroupTransform]):
-    """Mapping of destination group names to transform pipelines."""
-
-    def __init__(
-        self,
-        groups_dest: dict[str, GroupTransform] = {"default|Labels": GroupTransform()},
-    ):
-        super().__init__(groups_dest)
-
-
-class GroupMetric(dict[str, GroupTransformMetric]):
-    """Metric-oriented variant of :class:`Group` used during evaluation."""
-
-    def __init__(
-        self,
-        groups_dest: dict[str, GroupTransformMetric] = {"default|group_dest": GroupTransformMetric()},
-    ):
-        super().__init__(groups_dest)
-
-
-class GroupOut(dict[str, GroupTransformOut]):
-    """Transform-workflow variant of :class:`Group`."""
-
-    def __init__(
-        self,
-        groups_dest: dict[str, GroupTransformOut] = {"default|group_dest": GroupTransformOut()},
-    ):
-        super().__init__(groups_dest)
-
-
-def _chains(groups_src: Mapping[str, Group | GroupMetric | GroupOut]) -> Iterator[tuple[str, str, GroupTransform]]:
-    """Every ``(group_src, group_dest, chain)`` of the config, in declaration order."""
-    for group_src, group in groups_src.items():
-        for group_dest, chain in group.items():
-            yield group_src, group_dest, chain
-
-
-def _interleaved_case_entries(patches: list["DatasetPatch"], entries: list[tuple[int, int]]) -> list[tuple[int, int]]:
-    """One case's ``(copy, patch)`` entries ordered so the copies advance together along the slab axis.
-
-    A streamed TTA write reduces the copies slab by slab, so it can only advance to the slowest
-    copy's frontier: walked copy-major, the first copy would be complete (and fully retained)
-    before the second began. Ordering by each patch's declared first-spatial-axis start bounds that
-    skew at one patch extent, whatever grid each copy was cut on. The sort is total on
-    ``(start, copy, patch)``, so within a copy the order is untouched: per-copy accumulation is
-    byte-identical either way, and the whole-volume path reduces at the end whatever the order.
-
-    ``patches`` holds every destination group's grid for the case: one shared order must serve them
-    all, so if the groups disagree on the slab starts (or a group cannot even index an entry) the
-    plain order is kept. The interleave is a memory bound, never a correctness requirement.
-    """
-
-    def starts(patch: "DatasetPatch") -> list[int] | None:
-        try:
-            return [patch.get_patch_slices(copy)[index][0].start for copy, index in entries]
-        except (IndexError, KeyError):
-            return None
-
-    reference = starts(patches[0])
-    if reference is None or any(starts(patch) != reference for patch in patches[1:]):
-        return entries
-    order = dict(zip(entries, reference, strict=True))
-    return sorted(entries, key=lambda entry: (order[entry], *entry))
-
-
-class PatchReadOrder:
-    """The epoch's patch order, published from where it is drawn to where the patches are read.
-
-    A store that caches decoded chunks decides what to keep from the sequence it is told is coming
-    (:meth:`~konfai.utils.dataset.Dataset.plan_region_reads`). On the patch route that sequence is
-    the sampler's, drawn at the epoch's start in the parent, while the reads happen in the
-    DataLoader's workers, each holding a cache of its own: a worker is forked before the draw, and a
-    persistent one is never forked again, so the order reaches it through shared memory or not at
-    all. Nothing but the order travels: neither what is read nor its values depend on this.
-
-    A worker is handed one batch in ``num_workers``, so what it declares for a case is the
-    subsequence of its own batches. Their stride is learnt from the first batch it is given, which
-    assumes only that the batches are dealt round-robin, not which worker opens the epoch.
-    """
-
-    def __init__(self, mapping: list[tuple[int, int, int]], batch_size: int) -> None:
-        self._mapping = mapping
-        self._batch_size = max(1, batch_size)
-        self._order = torch.zeros(len(mapping), dtype=torch.int64).share_memory_()
-        self._epoch = torch.zeros(1, dtype=torch.int64).share_memory_()
-        self._epoch_read = 0
-        self._entries: dict[int, list[tuple[int, int]]] = {}
-
-    def publish(self, order: torch.Tensor) -> None:
-        """The order the epoch is about to be walked in, from the sampler that drew it."""
-        self._order.copy_(order)
-        self._epoch += 1
-
-    def entering(self, index: int) -> list[tuple[int, int]] | None:
-        """The ``(copy, patch)`` entries of ``index``'s case this process still has to read, in the
-        order it will read them; ``None`` once the case has been entered this epoch, and while no
-        sampler has published an order at all."""
-        self._regroup(index)
-        return self._entries.pop(self._mapping[index][0], None)
-
-    def _regroup(self, index: int) -> None:
-        """Group this process's remaining reads by case, once per epoch, from the batch ``index``
-        opens: a batch is ``batch_size`` consecutive positions, and the next batch this process is
-        handed is ``num_workers`` batches on."""
-        epoch = int(self._epoch[0])
-        if epoch == self._epoch_read:
-            return
-        self._epoch_read = epoch
-        order = self._order.tolist()
-        worker = data.get_worker_info()
-        stride = self._batch_size * (worker.num_workers if worker is not None else 1)
-        entries: dict[int, list[tuple[int, int]]] = {}
-        for start in range(order.index(index), len(order), stride):
-            for position in range(start, min(start + self._batch_size, len(order))):
-                case, copy, patch = self._mapping[order[position]]
-                entries.setdefault(case, []).append((copy, patch))
-        self._entries = entries
-
-
-class WindowedCaseSampler(Sampler[int]):
-    """Locality-aware training order: shuffle cases, window them, shuffle patches within each window.
-
-    ``DatasetIter`` loads each non-streamable case into a FIFO buffer, so a global patch shuffle
-    reloads a volume repeatedly (once per patch that lands after an eviction. Keeping only
-    ``window`` cases in play at a time) their patches shuffled together, emitted before advancing: reads each
-    volume ~once. ``window`` is the decorrelation knob: ``1`` is perfect locality, and
-    ``None`` (default) or ``>= n_cases`` is a single all-cases window, i.e. a plain global shuffle,
-    byte for byte.
-
-    A map-style ``DataLoader`` sends batch ``j`` to worker ``j % num_workers`` and gives each worker its
-    own buffer, so the cases are partitioned across workers (greedy least-loaded by patch count, see
-    ``_partitions``) and the per-worker windowed batches are round-robin interleaved: batch ``j`` then
-    carries only worker ``j % num_workers``'s cases, and every volume is read by exactly one worker.
-    """
-
-    def __init__(
-        self,
-        mapping: list[tuple[int, int, int]],
-        shuffle: bool,
-        window: int | None,
-        batch_size: int,
-        num_workers: int,
-        read_order: PatchReadOrder | None = None,
-    ) -> None:
-        self.mapping = mapping
-        self.shuffle = shuffle
-        # Where the epoch's order is published for the processes that read the patches; a sampler
-        # asked for nothing but its order publishes nowhere.
-        self.read_order = read_order
-        self.batch_size = max(1, batch_size)
-        self.num_workers = max(1, num_workers)
-        self.case_entries: dict[int, list[int]] = {}
-        for index, entry in enumerate(mapping):
-            self.case_entries.setdefault(entry[0], []).append(index)
-        # Windowing bites only when a window is both smaller than the case count and shardable across
-        # the workers; anything else is a single all-cases window, i.e. a plain global shuffle.
-        n_cases = len(self.case_entries)
-        self.window = window if window is not None and 0 < window < n_cases and self.num_workers <= n_cases else None
-
-    def _partitions(self) -> list[list[int]]:
-        """The cases each worker walks, balanced by the patches they hold.
-
-        A worker is handed whole cases, because a case is what its buffer keeps resident. Handing out
-        an equal COUNT of them leaves the patch counts as uneven as the cases are, and it is patches
-        that are walked: the workers then run out at different times, and the batches of whoever is
-        left shift onto the workers that finished: a case landing on two of them, each reading the
-        volume. Give the next case to whoever holds the fewest patches so far, largest first.
-        """
-        loads = [0] * self.num_workers
-        partitions: list[list[int]] = [[] for _ in range(self.num_workers)]
-        for case in sorted(self.case_entries, key=lambda case: -len(self.case_entries[case])):
-            worker = min(range(self.num_workers), key=lambda worker: loads[worker])
-            partitions[worker].append(case)
-            loads[worker] += len(self.case_entries[case])
-        return partitions
-
-    def _windowed_order(self) -> list[int]:
-        generator = torch.Generator().manual_seed(int(torch.randint(0, 2**31 - 1, (1,)).item()))
-        window = cast(int, self.window)
-
-        def shuffled(items: list[int]) -> list[int]:
-            return [items[i] for i in torch.randperm(len(items), generator=generator).tolist()]
-
-        streams: list[list[int]] = []
-        for partition in self._partitions():
-            cases = shuffled(partition)
-            stream: list[int] = []
-            for start in range(0, len(cases), window):
-                stream += shuffled(
-                    [index for case in cases[start : start + window] for index in self.case_entries[case]]
-                )
-            streams.append(stream)
-
-        # Round-robin the workers a batch at a time, so batch j lands on worker j % num_workers and a
-        # window stays resident while its patches are walked.
-        #
-        # Three things are wanted here and only two of them fit. An epoch must be one pass over the
-        # mapping; a worker must keep whole cases, since a case is what its buffer holds; and a case
-        # should stay on one worker, which needs every stream the same length. A case does not split,
-        # so streams of equal length are not something `_partitions` can always hand over, one case
-        # of 200 patches beside ten of 2 is longer on its own than a quarter of the epoch.
-        #
-        # So a short stream runs out and the ones still going shift onto the workers that finished:
-        # a case lands on two of them and each reads its volume. The epoch stays exact and the reads
-        # stay close to 1x: far below the redundant reads of no window at all. Padding the streams
-        # instead buys the affinity back by walking part of the epoch twice and the rest not at all,
-        # which is not a trade to make.
-        batch = self.batch_size
-        order: list[int] = []
-        for start in range(0, max((len(stream) for stream in streams), default=0), batch):
-            for stream in streams:
-                order += stream[start : start + batch]
-        return order
-
-    def __iter__(self) -> Iterator[int]:
-        if not self.shuffle:
-            order = torch.arange(len(self.mapping))
-        elif self.window is None:
-            order = torch.randperm(len(self.mapping))
-        else:
-            order = torch.as_tensor(self._windowed_order(), dtype=torch.int64)
-        if self.read_order is not None:
-            self.read_order.publish(order)
-        return iter(order.tolist())
-
-    def __len__(self) -> int:
-        # One epoch is one pass over the mapping: windowing chooses the ORDER, not the size. This is
-        # what keeps the ranks in step: `Data._split` gives them equal-length shards, but not equal
-        # cases, so any length read from the per-rank cases (their partitions, or even whether the
-        # window engages at all) would differ and hang DDP's collectives.
-        return len(self.mapping)
-
-
-@dataclass(frozen=True)
-class DataItem:
-    """Single tensor sample together with dataset metadata and patch indices."""
-
-    name: str
-    tensor: torch.Tensor
-    attribute: Attribute
-    x: int
-    a: int
-    p: int
-    is_input: bool
-
-
-@dataclass(frozen=True)
-class BatchDataItem:
-    """Batch-level representation of multiple :class:`DataItem` objects."""
-
-    name: list[str]
-    tensor: torch.Tensor  # [B, ...]
-    attribute: list[Attribute]
-    x: list[int]
-    a: list[int]
-    p: list[int]
-    is_input: bool
-
-    def pin_memory(self) -> "BatchDataItem":
-        """The batch with its tensor in page-locked memory, so the upload is a real DMA.
-
-        ``torch``'s pinner walks tensors, mappings and sequences and hands anything else back
-        untouched: without this method ``pin_memory: true`` reaches nothing and every upload
-        stays a pageable copy the host has to wait for.
-        """
-        return replace(self, tensor=self.tensor.pin_memory())
-
-
-Sample: TypeAlias = dict[str, DataItem]
-BatchSample: TypeAlias = dict[str, BatchDataItem]
-
-
-def collate_konfai(batch: list[Sample]) -> BatchSample:
-    """Collate KonfAI samples into the batch structure expected by the workflows."""
-    batch_sample: BatchSample = {}
-    for k in batch[0].keys():
-        items = [b[k] for b in batch]
-        batch_sample[k] = BatchDataItem(
-            tensor=torch.stack([it.tensor for it in items], dim=0),
-            x=[it.x for it in items],
-            a=[it.a for it in items],
-            p=[it.p for it in items],
-            attribute=[it.attribute for it in items],
-            name=[it.name for it in items],
-            is_input=items[0].is_input,
-        )
-    return batch_sample
-
-
-class DatasetIter(data.Dataset):
-    """Torch dataset view over KonfAI dataset managers and patch mappings."""
-
-    def __init__(
-        self,
-        rank: int,
-        data: dict[str, list[DatasetManager]],
-        mapping: list[tuple[int, int, int]],
-        groups_src: Mapping[str, Group | GroupMetric | GroupOut],
-        inline_augmentations: bool,
-        data_augmentations_list: list[DataAugmentationsList],
-        patch_size: list[int] | None,
-        overlap: OverlapSpec,
-        buffer_size: int,
-        apply_augmentations: bool = True,
-        use_cache=True,
-        batch_size: int = 1,
-    ) -> None:
-        self.rank = rank
-        self.data = data
-        self.mapping = mapping
-        self.patch_size = patch_size
-        self.overlap = overlap
-        self.groups_src = groups_src
-        self.apply_augmentations = apply_augmentations
-        self.data_augmentations_list = data_augmentations_list if apply_augmentations else []
-        self.use_cache = use_cache
-        self.nb_dataset = len(data[next(iter(data.keys()))])
-        self.buffer_size = buffer_size
-        self._index_cache: list[int] = []
-        self._index_cache_lookup: set[int] = set()
-        self.inline_augmentations = inline_augmentations
-        self.has_augmented_samples = self.apply_augmentations and any(a > 0 for _, a, _ in mapping)
-        self.read_order = PatchReadOrder(mapping, batch_size)
-
-    def get_patch_config(self) -> tuple[list[int] | None, OverlapSpec]:
-        return self.patch_size, self.overlap
-
-    def to(self, device: int):
-        for _group_src, _group_dest, chain in _chains(self.groups_src):
-            chain.to(device)
-        for data_augmentations in self.data_augmentations_list:
-            for data_augmentation in data_augmentations.data_augmentations:
-                data_augmentation.to(device)
-
-    def get_dataset_from_index(self, group_dest: str, index: int) -> DatasetManager:
-        return self.data[group_dest][index]
-
-    def reset_augmentation(self, label):
-        if self.inline_augmentations and self.has_augmented_samples and len(self.data_augmentations_list) > 0:
-            for index in range(self.nb_dataset):
-                # Augmentation objects are shared across destination groups AND across the train and
-                # validation loaders, so the per-case draw is cached by the manager's own augmentation
-                # index (globally unique, offset for validation), not the loader-local position: else a
-                # validation case would reset (and reuse) a train case's draw and folded shape.
-                case_index = next(iter(self.data.values()))[index].index
-                for data_augmentations in self.data_augmentations_list:
-                    for data_augmentation in data_augmentations.data_augmentations:
-                        data_augmentation.reset_state(case_index)
-                for _group_src, group_dest, _chain in _chains(self.groups_src):
-                    self.data[group_dest][index].unload_augmentation()
-                    self.data[group_dest][index].reset_augmentation(reset_state=False)
-            self.load(label + " Augmentation")
-
-    def load(self, label: str):
-        if self.use_cache:
-            memory_init = get_memory()
-
-            indexs = list(range(self.nb_dataset))
-            if len(indexs) > 0:
-                memory_lock = threading.Lock()
-
-                def desc(i: int = 0):
-                    return (
-                        f"Caching {label}: "
-                        f"{get_memory_info()} | "
-                        f"{memory_forecast(memory_init, i, self.nb_dataset)} | "
-                        f"{get_cpu_info()}"
-                    )
-
-                pbar = tqdm.tqdm(total=len(indexs), desc=desc(), leave=False)
-                stop_event = threading.Event()
-
-                def process(index):
-                    if stop_event.is_set():
-                        return
-                    self._load_data(index)
-                    with memory_lock:
-                        pbar.set_description(desc(pbar.n + 1))
-                        pbar.update(1)
-
-                cpu_count = os.cpu_count() or 1
-                try:
-                    with ThreadPoolExecutor(max_workers=_cache_worker_count(cpu_count, device_count())) as executor:
-                        future_to_index = {executor.submit(process, index): index for index in indexs}
-                        for fut in as_completed(future_to_index):
-                            index = future_to_index[fut]
-                            try:
-                                fut.result()
-                            except Exception as e:
-                                stop_event.set()
-                                for f in future_to_index:
-                                    f.cancel()
-                                tb = traceback.format_exc()
-                                raise RuntimeError(
-                                    f"Error while caching {label} (index={index})\n"
-                                    f"{type(e).__name__}: {e}\n\n"
-                                    f"Traceback (worker):\n{tb}"
-                                ) from e
-
-                except KeyboardInterrupt:
-                    stop_event.set()
-                    try:
-                        for f in future_to_index:
-                            f.cancel()
-                    except Exception:  # nosec B110
-                        pass
-                    raise
-                finally:
-                    pbar.close()
-
-    def _load_data(self, index: int, augmentation_index: int | None = None) -> bool:
-        loaded = False
-        for group_src, group_dest, _chain in _chains(self.groups_src):
-            loaded |= self.load_data(group_src, group_dest, index, augmentation_index)
-        if loaded and index not in self._index_cache_lookup:
-            self._index_cache.append(index)
-            self._index_cache_lookup.add(index)
-        return loaded
-
-    def load_data(self, group_src: str, group_dest: str, index: int, augmentation_index: int | None = None) -> bool:
-        item = self.data[group_dest][index]
-        if augmentation_index is not None and item.can_stream_patch(augmentation_index, self.apply_augmentations):
-            return False
-        try:
-            item.load(
-                self.groups_src[group_src][group_dest].transforms,
-                self.data_augmentations_list,
-                load_augmentations=self.apply_augmentations and not self.inline_augmentations,
-            )
-        except Exception as e:
-            raise RuntimeError(
-                f"Error while loading data "
-                f"(group_src={group_src}, group_dest={group_dest}, "
-                f"index={index}, name={item.name}) : "
-                f"{type(e).__name__}: {e}"
-            ) from e
-        return True
-
-    def _unload_data(self, index: int) -> None:
-        if index in self._index_cache_lookup:
-            self._index_cache_lookup.remove(index)
-            self._index_cache.remove(index)
-        for _group_src, group_dest, _chain in _chains(self.groups_src):
-            self.unload_data(group_dest, index)
-
-    def unload_data(self, group_dest: str, index: int) -> None:
-        return self.data[group_dest][index].unload()
-
-    def _declare_case_reads(self, index: int) -> None:
-        """Tell each group's store the patches this process will read of the case ``index`` enters,
-        in the order it will read them: once per case, at the first patch of it that arrives."""
-        entries = self.read_order.entering(index)
-        if entries is None:
-            return
-        case = self.mapping[index][0]
-        for _group_src, group_dest, chain in _chains(self.groups_src):
-            self.data[group_dest][case].plan_patch_reads(entries, chain.is_input, self.apply_augmentations)
-
-    def __len__(self) -> int:
-        return len(self.mapping)
-
-    def __getitem__(self, index: int) -> Sample:
-        sample: Sample = {}
-        x, a, p = self.mapping[index]
-        needs_full_load = any(
-            not self.data[group_dest][x].can_stream_patch(a, self.apply_augmentations)
-            for _group_src, group_dest, _chain in _chains(self.groups_src)
-        )
-        if x not in self._index_cache_lookup and needs_full_load:
-            if len(self._index_cache) >= self.buffer_size and not self.use_cache:
-                self._unload_data(self._index_cache[0])
-            self._load_data(x, a)
-
-        self._declare_case_reads(index)
-
-        for _group_src, group_dest, chain in _chains(self.groups_src):
-            dataset = self.data[group_dest][x]
-            sample[f"{group_dest}"] = DataItem(
-                dataset.name,
-                dataset.get_data(p, a, chain.patch_transforms, chain.is_input, self.apply_augmentations),
-                dataset.cache_attributes[a],
-                x,
-                a,
-                p,
-                chain.is_input,
-            )
-        return sample
-
-
-class Subset:
-    def __init__(
-        self,
-        subset: str | list[int] | list[str] | None = None,
-        shuffle: bool = True,
-        shuffle_window: int | None = None,
-    ) -> None:
-        self.subset = subset
-        self.shuffle = shuffle
-        self.shuffle_window = shuffle_window
-
-    @staticmethod
-    def _read_names_from_file(filename: str) -> list[str]:
-        with open(filename) as f:
-            return [name.strip() for name in f if name.strip()]
-
-    def requires_infos(self) -> bool:
-        """Return whether this subset implementation needs per-sample metadata."""
-        return self.__class__.__call__ is not Subset.__call__
-
-    @staticmethod
-    def _is_slice_selector(subset: str) -> bool:
-        start, sep, end = subset.partition(":")
-        if sep == "":
-            return False
-        return start.lstrip("-").isdigit() and end.lstrip("-").isdigit()
-
-    def _resolve_selector(self, subset: str | int, names: list[str]) -> tuple[set[int], bool]:
-        size = len(names)
-        name_to_index = {name: i for i, name in enumerate(names)}
-
-        if isinstance(subset, int):
-            return {subset}, False
-        if subset.startswith("~"):
-            excluded = subset[1:]
-            if os.path.exists(excluded):
-                exclude_names = set(self._read_names_from_file(excluded))
-                return {i for i, name in enumerate(names) if name in exclude_names}, True
-            if excluded in name_to_index:
-                return {name_to_index[excluded]}, True
-            return set(), True
-        if os.path.exists(subset):
-            selected_names = set(self._read_names_from_file(subset))
-            return {i for i, name in enumerate(names) if name in selected_names}, False
-        if self._is_slice_selector(subset):
-            start, _, end = subset.partition(":")
-            r = np.clip(
-                np.asarray([int(start), int(end)]),
-                0,
-                size,
-            )
-            return set(range(int(r[0]), int(r[1]))), False
-        if subset in name_to_index:
-            return {name_to_index[subset]}, False
-        return set(), False
-
-    def _get_index(self, subset: str | int, names: list[str]) -> list[int]:
-        index, is_exclusion = self._resolve_selector(subset, names)
-        if is_exclusion:
-            return [i for i in range(len(names)) if i not in index]
-        return sorted(index)
-
-    @staticmethod
-    def _excludes(selector: str | int) -> bool:
-        return isinstance(selector, str) and selector.startswith("~")
-
-    def _included_names(self, selector: str | int) -> set[str] | None:
-        """The cases ``selector`` keeps, or ``None`` when only the cohort can say which they are.
-        Read in ``_resolve_selector``'s order, a file before a slice, so the walk asks for what the
-        selection then keeps."""
-        if isinstance(selector, int):
-            return None  # a position, and positions are defined against the full sorted list
-        if os.path.exists(selector):
-            return set(self._read_names_from_file(selector))
-        if self._is_slice_selector(selector):
-            return None
-        return {selector}
-
-    def required_names(self) -> set[str] | None:
-        """The cases this subset keeps, when it can name them, or ``None`` when it needs the cohort.
-
-        A root is asked for these instead of for everything it holds. Exclusions do not appear:
-        they only remove from what the inclusions brought, and a subset that ONLY excludes is
-        defined against the full list, so it asks for it.
-        """
-        if self.requires_infos():
-            return None  # a subclass selecting on geometry: it is handed every case, and picks
-        selectors = self.subset if isinstance(self.subset, list) else [self.subset]
-        included = [self._included_names(s) for s in selectors if s is not None and not self._excludes(s)]
-        return None if not included or None in included else set().union(*included)  # type: ignore[arg-type]
-
-    def __call__(self, names: list[str], infos: dict[str, tuple[list[int], Attribute]]) -> set[str]:
-        names = sorted(names)
-        size = len(names)
-
-        if self.subset is None:
-            index = list(range(0, size))
-        elif isinstance(self.subset, list):
-            if len(self.subset) == 0:
-                index = []
-            else:
-                include_index: set[int] = set()
-                exclude_index: set[int] = set()
-                has_include = False
-                for s in self.subset:
-                    resolved_index, is_exclusion = self._resolve_selector(s, names)
-                    if is_exclusion:
-                        exclude_index.update(resolved_index)
-                    else:
-                        include_index.update(resolved_index)
-                        has_include = True
-                index_set = include_index if has_include else set(range(size))
-                index = sorted(index_set.difference(exclude_index))
-        else:
-            index = self._get_index(self.subset, names)
-        return {names[i] for i in index}
-
-    def __str__(self):
-        return f"Subset : {self.subset} shuffle : {self.shuffle} shuffle_window : {self.shuffle_window}"
-
-
-class TrainSubset(Subset):
-    def __init__(
-        self,
-        subset: str | list[int] | list[str] | None = None,
-        shuffle: bool = True,
-        shuffle_window: int | None = None,
-    ) -> None:
-        super().__init__(subset, shuffle, shuffle_window)
-
-
-class PredictionSubset(Subset):
-    def __init__(self, subset: str | list[int] | list[str] | None = None) -> None:
-        super().__init__(subset, False, None)
+from konfai.utils.errors import DatasetManagerError, TransformerError
+from konfai.utils.runtime import State
+from konfai.utils.utils import SUPPORTED_FORMATS, resolve_patch, split_path_spec
 
 
 class DataSources(ABC):
