+++ konfai/predictor/output.py
@@ -14,36 +14,24 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""Prediction workflow classes and the streamed-write dispatcher."""
 
-import copy
+"""The prediction sink: patches blended per case and copy, reduced, written whole or streamed by slabs."""
+
 import os
 import queue
-import shutil
 import threading
 import warnings
-from collections import defaultdict
-from collections.abc import Callable, Mapping
+from collections.abc import Callable
 from contextlib import suppress
 from dataclasses import dataclass
-from pathlib import Path
-from typing import Any, cast
+from typing import cast
 
 import numpy as np
 import torch
-import tqdm
-from torch.utils.data import DataLoader
 
-try:
-    from torch.utils.tensorboard.writer import SummaryWriter
-except ImportError:
-    SummaryWriter = None  # type: ignore[assignment,misc]
-
-from konfai import config_file, cuda_visible_devices, konfai_root, predictions_directory
+from konfai import konfai_root
 from konfai.data.augmentation import DataAugmentation
 from konfai.data.data_manager import (
-    BatchSample,
-    DataPrediction,
     DatasetIter,
 )
 from konfai.data.patching import (
@@ -58,11 +46,7 @@ from konfai.data.patching import (
     blend_axes,
     blend_overlap,
 )
-
-# Re-exported: the operators moved to konfai.data.reduction, where the transform workflow can
-# reach them without importing the predictor. Existing configs name them by the classpath
-# `konfai.predictor.Median`, so the names must stay resolvable here.
-from konfai.data.reduction import Concat, Mean, Median, Reduction
+from konfai.data.reduction import Mean, Median, Reduction
 from konfai.data.transform import (
     LocalityKind,
     PatchLocality,
@@ -71,35 +55,19 @@ from konfai.data.transform import (
     TransformInverse,
     TransformLoader,
 )
-from konfai.network.network import Model, ModelLoader, NetState, Network
-from konfai.utils import vram
-from konfai.utils.budget import node_local_ranks, resolve_memory_budget, set_per_rank_budget
-from konfai.utils.chain_diff import dataset_tree, input_chain_differences, training_dataset_tree
-from konfai.utils.clock import SweepClock, startup_clock
-from konfai.utils.config import _escape_key_component, apply_config, config, strict_config
+from konfai.utils.budget import node_local_ranks, resolve_memory_budget
+from konfai.utils.clock import SweepClock
+from konfai.utils.config import _escape_key_component, apply_config, config
 from konfai.utils.dataset import Attribute, Dataset, DataStream
-from konfai.utils.errors import ConfigError, KonfAIError, PredictorError
-from konfai.utils.ome_zarr import bound_chunk_cache
+from konfai.utils.errors import PredictorError
 from konfai.utils.runtime import (
-    DataLog,
-    DistributedObject,
     NeedDevice,
-    State,
-    configure_workflow_environment,
-    confirm_overwrite_or_raise,
-    description,
-    run_distributed_app,
-    safe_torch_load,
 )
-from konfai.utils.utils import concretize_patch_size, env_flag, get_module, size_free_axes, split_path_spec
+from konfai.utils.utils import env_flag, get_module, split_path_spec
 
 #: This rank's prediction loop, phase by phase, summed over the cases it ran: the unit the line
 #: at the end of a run accounts for (see ``_prediction_report``).
 PREDICTION_CLOCK = SweepClock()
-
-#: Batches between two refreshes of the progress bar's status. The status is an NVML query, two
-#: psutil calls and a forced redraw (95 us per batch, measured) that no batch needs to be current.
-_DESCRIPTION_EVERY = 10
 
 
 class _AsyncWriter:
@@ -148,14 +116,6 @@ class _AsyncWriter:
         if self._error is not None:
             error, self._error = self._error, None
             raise error
-
-
-# Streaming pays per-slab work (the pipe traversal, region writes, the TTA aligner); when the
-# assembled accumulators of all copies are below this fraction of allocatable memory (a 2.5D case),
-# holding them whole costs nothing and the case takes the whole-volume path.
-# KONFAI_STREAM_WORTH_THRESHOLD overrides the fraction (tests set 0 to exercise the streamed
-# machinery on toy volumes). See OutputDataset._worth_streaming.
-_STREAM_WORTH_MIN_FRACTION = 0.05
 
 
 def _slab_context(region: slice, spatial: list[int]) -> RegionContext:
@@ -239,6 +199,14 @@ class _RegionState:
     stream: SlabRegionStream | None = None
     # The attribute the latest emission ran the pipe on: what the sink opens with.
     attribute: Attribute | None = None
+
+
+# Streaming pays per-slab work (the pipe traversal, region writes, the TTA aligner); when the
+# assembled accumulators of all copies are below this fraction of allocatable memory (a 2.5D case),
+# holding them whole costs nothing and the case takes the whole-volume path.
+# KONFAI_STREAM_WORTH_THRESHOLD overrides the fraction (tests set 0 to exercise the streamed
+# machinery on toy volumes). See OutputDataset._worth_streaming.
+_STREAM_WORTH_MIN_FRACTION = 0.05
 
 
 @config("OutputDataset")
@@ -1439,940 +1407,3 @@ class OutputDatasetLoader:
     def get_output_dataset(self, layer_name: str) -> OutputDataset:
         module, name = get_module(self.name_class, "konfai.predictor")
         return apply_config(f"Predictor.outputs_dataset.{layer_name}")(getattr(module, name))()
-
-
-def _prediction_report(clock: SweepClock, min_seconds: float = 1.0) -> str | None:
-    """One line accounting for the prediction loop's wall clock, in the sweep report's shape, or
-    ``None`` below ``min_seconds``.
-
-    The sum before the bar is the loop's own thread and closes exactly: what its phases do not
-    name (the logging, the progress bar, the bookkeeping between them) is ``other``. ``fetch`` is
-    the wait for the loader's next batch, ``blend`` a patch's inverses and its blend into the
-    accumulator (the copy home included on the host route), the two ``finalize`` the slabs and the
-    cases handed to the writer, ``drain`` the writes still queued when the loop ends. On a GPU the
-    loop only enqueues the forward and the blend; the device's time is waited for where a result
-    crosses to the host.
-
-    After the bar is the writer: its own thread's time, and how long the loop stood waiting on it
-    inside the finalize phases (a full queue, or the write itself when the destination keeps it
-    inline). A slow destination shows there, as the wait that turns the background writer
-    synchronous, which nothing reported before.
-    """
-    wall = clock.spent("prediction")
-    if wall < min_seconds:
-        return None
-    phases = ("fetch", "forward", "blend", "finalize(stream)", "finalize(case)", "drain")
-    named = {phase: clock.spent(phase) for phase in phases}
-    parts = " + ".join(f"{phase} {value:.1f}" for phase, value in named.items())
-    return (
-        f"[KonfAI] prediction {wall:.1f} s = {parts} + other {wall - sum(named.values()):.1f}"
-        f" | writer {clock.spent('write'):.1f} s, waited on {clock.spent('wait(write)'):.1f} s"
-    )
-
-
-class _Predictor:
-    """
-    Internal class that runs distributed inference over a dataset using a composite model.
-
-    This class handles patch-wise prediction, output accumulation, logging to TensorBoard, and
-    writing final predictions to disk. It is designed to be used as a context manager and
-    supports model ensembles via `ModelComposite`.
-
-    Args:
-        world_size (int): Total number of processes or GPUs used.
-        global_rank (int): Rank of the current process across all nodes.
-        local_rank (int): Local GPU index within a single node.
-        autocast (bool): Whether to use automatic mixed precision (AMP).
-        predict_path (str): Output directory path where predictions and metrics are saved.
-        data_log (list[str] | None): List of logging targets in the format 'group/DataLogType/N'.
-        outputs_dataset (dict[str, OutputDataset]): Dictionary of output datasets to store predictions.
-        model_composite (Model): Model container that wraps the prediction model(s).
-        dataloader_prediction (DataLoader): DataLoader that provides prediction batches.
-    """
-
-    def __init__(
-        self,
-        world_size: int,
-        global_rank: int,
-        local_rank: int,
-        autocast: bool,
-        predict_path: Path,
-        data_log: list[str] | None,
-        outputs_dataset: dict[str, OutputDataset],
-        model_composite: Model,
-        dataloader_prediction: DataLoader,
-    ) -> None:
-        self.world_size = world_size
-        self.global_rank = global_rank
-        self.local_rank = local_rank
-
-        self.model_composite = model_composite
-
-        self.dataloader_prediction = dataloader_prediction
-        self.outputs_dataset = outputs_dataset
-        self.autocast = autocast
-        self.it = 0
-
-        self.dataset: DatasetIter = self.dataloader_prediction.dataset
-        patch_size, overlap = self.dataset.get_patch_config()
-        for output_dataset in self.outputs_dataset.values():
-            output_dataset.set_patch_config(
-                patch_size,
-                overlap,
-                np.max(
-                    [
-                        int(
-                            np.sum([data_augmentation.nb for data_augmentation in self.dataset.data_augmentations_list])
-                            + 1
-                        ),
-                        1,
-                    ]
-                ),
-            )
-        self.data_log = DataLog.parse(data_log)
-        self._has_runtime_measures = any(
-            network.measure is not None for network in self.model_composite.module.get_networks().values()
-        )
-        if self._has_runtime_measures or len(self.data_log):
-            if SummaryWriter is None:
-                raise ImportError(
-                    "TensorBoard is required for prediction logging. Install it with: pip install konfai[tensorboard]"
-                )
-            self.tb = SummaryWriter(log_dir=predict_path / "Metric")
-        else:
-            self.tb = None
-
-    def __enter__(self):
-        """
-        Enters the prediction context and returns the predictor instance.
-        """
-        return self
-
-    def __exit__(self, exc_type, value, traceback):
-        """
-        Closes the TensorBoard writer upon exit.
-        """
-        if self.tb:
-            self.tb.close()
-
-    def run(self):
-        """
-        Run the full prediction loop.
-
-        Iterates over the prediction DataLoader, performs inference using the composite model,
-        applies reduction (e.g., mean), and writes the final results using each `OutputDataset`.
-
-        Also logs intermediate data and metrics to TensorBoard if enabled.
-        """
-
-        self.model_composite.eval()
-        self.model_composite.module.set_state(NetState.PREDICTION)
-        self.dataloader_prediction.dataset.load("Prediction")
-        PREDICTION_CLOCK.reset()
-        try:
-            with PREDICTION_CLOCK.phase("prediction"):
-                self._run_batches()
-        finally:
-            # Every submitted write must be on disk before the run returns: including on the error
-            # path, where the drain also closes the sinks the abort operations enqueued.
-            with PREDICTION_CLOCK.phase("prediction"), PREDICTION_CLOCK.phase("drain"):
-                for output_dataset in self.outputs_dataset.values():
-                    output_dataset.finalize_writes()
-        if self.global_rank == 0:
-            clock = _prediction_report(PREDICTION_CLOCK)
-            if clock is not None:
-                print(clock)
-
-    def _run_batches(self) -> None:
-        with tqdm.tqdm(
-            iterable=enumerate(PREDICTION_CLOCK.waiting("fetch", self.dataloader_prediction)),
-            leave=True,
-            desc=f"Prediction : {description(self.model_composite)}",
-            total=len(self.dataloader_prediction),
-            ncols=0,
-        ) as batch_iter:
-            with torch.inference_mode():
-                with torch.amp.autocast("cuda", enabled=self.autocast):
-                    for batch_index, batch_sample in batch_iter:
-                        with PREDICTION_CLOCK.phase("forward"):
-                            outputs = self.model_composite(
-                                batch_sample,
-                                list(self.outputs_dataset.keys()),
-                            )
-                        self._predict_log(batch_sample)
-                        for name, number_of_channels_per_model, output in outputs:
-                            output_dataset = self.outputs_dataset[name]
-                            group = getattr(output_dataset, "group_dest", next(iter(batch_sample)))
-                            for i, (index, patch_augmentation, patch_index) in enumerate(
-                                [
-                                    (int(index), int(patch_augmentation), int(patch_index))
-                                    for index, patch_augmentation, patch_index in zip(
-                                        batch_sample[group].x,
-                                        batch_sample[group].a,
-                                        batch_sample[group].p,
-                                        strict=False,
-                                    )
-                                ]
-                            ):
-                                output_dataset.add_layer(
-                                    index,
-                                    patch_augmentation,
-                                    patch_index,
-                                    output[i],
-                                    self.dataset,
-                                    batch_sample[group].attribute[i],
-                                    number_of_channels_per_model,
-                                )
-                                if output_dataset.is_done(index):
-                                    with PREDICTION_CLOCK.phase("finalize(case)"):
-                                        output_dataset.write_prediction(
-                                            index,
-                                            batch_sample[group].name[i],
-                                            output_dataset.get_output(
-                                                index, number_of_channels_per_model, self.dataset
-                                            ),
-                                        )
-
-                        if batch_index % _DESCRIPTION_EVERY == 0:
-                            # The bar redraws on its own clock; the status only has to be there by then.
-                            batch_iter.set_description(
-                                f"Prediction : {description(self.model_composite)}", refresh=False
-                            )
-                        self.it += 1
-
-    def _predict_log(
-        self,
-        batch_sample: BatchSample,
-    ):
-        """
-        Log prediction results to TensorBoard, including images and metrics.
-
-        This method handles:
-        - Logging image-like data (e.g., inputs, outputs, masks) using `DataLog` instances,
-        based on the `data_log` configuration.
-        - Logging scalar loss and metric values (if present in the network) under the `Prediction/` namespace.
-        - Dynamically retrieving additional feature maps or intermediate layers if requested via `data_log`.
-
-        Logging is performed only on the global rank 0 process and only if `TensorBoard` is active.
-
-        Args:
-            data_dict (dict): Dictionary mapping group names to 6-tuples containing:
-                - input tensor,
-                - index,
-                - patch_augmentation,
-                - patch_index,
-                - metadata (list of strings),
-                - `requires_grad` flag (as a tensor).
-        """
-        if self.tb is None or self.global_rank != 0:
-            # Prediction logging is a rank-0 progress indicator; gate before touching the measures so a
-            # non-zero rank never enters a cross-rank collective the unequal shards would deadlock on.
-            return
-
-        measures: dict[str, tuple[dict[str, tuple[float, float]], dict[str, tuple[float, float]]]] = {}
-        if self._has_runtime_measures:
-            measures = DistributedObject.get_measure(
-                1,
-                0,
-                self.local_rank,
-                {"": self.model_composite.module},
-                1,
-                sync=False,
-            )
-
-        images_log = []
-        if len(self.data_log):
-            for name, data_type in self.data_log.items():
-                if name in batch_sample:
-                    data_type[0](
-                        self.tb,
-                        f"Prediction/{name}",
-                        batch_sample[name].tensor[: self.data_log[name][1]].detach().cpu().numpy(),
-                        self.it,
-                    )
-                else:
-                    images_log.append(name.replace(":", "."))
-
-        for name, network in self.model_composite.module.get_networks().items():
-            if network.measure is not None:
-                self.tb.add_scalars(
-                    f"Prediction/{name}/Loss",
-                    {k.replace(":", "."): v[1] for k, v in measures[name][0].items()},
-                    self.it,
-                )
-                self.tb.add_scalars(
-                    f"Prediction/{name}/Metric",
-                    {k.replace(":", "."): v[1] for k, v in measures[name][1].items()},
-                    self.it,
-                )
-            if len(images_log):
-                for name, layer, _ in self.model_composite.module.get_layers(
-                    [v.tensor for v in batch_sample.values() if v.is_input],
-                    images_log,
-                ):
-                    self.data_log[name][0](
-                        self.tb,
-                        f"Prediction/{name}",
-                        layer[: self.data_log[name][1]].detach().cpu().numpy(),
-                        self.it,
-                    )
-
-
-def _colocate_loaded_modules(model: torch.nn.Module) -> None:
-    """Move any still-CPU leaf module onto the model's device.
-
-    A custom :meth:`Network.load` may append modules after the model was already placed on its
-    device (e.g. a head sized from the checkpoint's class count), and those default to CPU, which
-    then raises a device mismatch on the forward pass. This re-homes any fully-CPU leaf onto the
-    device the rest of the model already lives on. Modules already on a device (including
-    model-parallel splits across several GPUs) are left untouched.
-    """
-    target = next((p.device for p in model.parameters() if p.device.type != "cpu"), None)
-    if target is None:
-        return
-    for sub in model.modules():
-        # ModuleArgsDict overrides parameters()/buffers() without a ``recurse`` kwarg, so use the
-        # base nn.Module methods to read each module's own (non-recursive) tensors.
-        own = [
-            *torch.nn.Module.parameters(sub, recurse=False),
-            *torch.nn.Module.buffers(sub, recurse=False),
-        ]
-        if own and all(t.device.type == "cpu" for t in own):
-            sub.to(target)
-
-
-class ModelComposite(Network):
-    """
-    A composite model that replicates a given base network multiple times and combines their outputs.
-
-    This class is designed to handle model ensembles or repeated predictions from the same architecture.
-    It creates `nb_models` deep copies of the input `model`, each with its own name and output branch,
-    and aggregates their outputs using a provided `Reduction` strategy (e.g., mean, median).
-
-    Args:
-        model (Network): The base network to replicate.
-        nb_models (int): Number of copies of the model to create.
-        combine (konfai.data.reduction.Reduction): The reduction method used to combine outputs from
-            all model replicas.
-
-    Attributes:
-        combine (konfai.data.reduction.Reduction): The reduction used during forward inference.
-    """
-
-    def __init__(self, model: Network, combine: Reduction):
-        super().__init__(
-            model.in_channels,
-            model.optimizer,
-            model.lr_schedulers_loader,
-            model.outputs_criterions_loader,
-            model.patch,
-            model.nb_batch_per_step,
-            model.init_type,
-            model.init_gain,
-            model.dim,
-        )
-        self.combine = combine
-        self._model_name = "Model_0"
-        self._base_model_name = model.get_name()
-        self._state_sources: list[dict[str, Any] | Path | str] = []
-        self._loaded = False  # load() has run: distinguishes "not loaded yet" from "loaded, weightless"
-        self._loaded_state_index: int | None = None
-        # Cache the CPU state_dict per index so a local-path ensemble is read from
-        # disk once, not re-read + re-unpickled on every batch (the index cycles
-        # 0..N-1 each forward, so the next batch would otherwise reload all N).
-        self._state_cache: dict[int, dict[str, Any]] = {}
-        self.add_module(
-            self._model_name,
-            copy.deepcopy(model),
-            in_branch=[0],
-            out_branch=["output_0"],
-        )
-
-    def _get_model(self) -> Network:
-        return cast(Network, self[self._model_name])
-
-    def _read_state_source(self, source: dict[str, Any] | Path | str) -> dict[str, Any]:
-        if isinstance(source, dict):
-            return source
-        if isinstance(source, str) and source.startswith("https://"):
-            return torch.hub.load_state_dict_from_url(url=source, map_location="cpu", check_hash=True)
-        return safe_torch_load(source, torch.device("cpu"))
-
-    def _ensure_model_loaded(self, index: int) -> Network:
-        model = self._get_model()
-        if self._loaded_state_index != index:
-            state = self._state_cache.get(index)
-            if state is None:
-                state = self._read_state_source(self._state_sources[index])
-                self._state_cache[index] = state
-            # Checkpoints are keyed by the base model name, not by the streamed
-            # ensemble suffix added after the previous load.
-            model.set_name(self._base_model_name)
-            model.load(state, init=False)
-            # A custom load() may append checkpoint-sized modules (e.g. the head) on CPU; co-locate
-            # them with the already device-placed model so the forward pass doesn't hit a mismatch.
-            _colocate_loaded_modules(model)
-            model.set_name(f"{self._base_model_name}_{index}")
-            self._loaded_state_index = index
-        return model
-
-    def _model_for_index(self, index: int) -> Network:
-        # With no checkpoint sources the model is weightless (0 parameters, e.g. a classical/optimisation
-        # engine): run it as constructed, once. The Predictor guards this, it only reaches here with empty
-        # sources when the model has no parameters to load, so there is nothing to stream.
-        if not self._state_sources:
-            return self._get_model()
-        return self._ensure_model_loaded(index)
-
-    def load(self, state_sources: list[dict[str, Any] | Path | str]):
-        """
-        Load weights for each sub-model in the composite from the corresponding state dictionaries.
-
-        Args:
-            state_sources (list): One checkpoint source per model replica. Empty ONLY for a weightless model
-                (0 parameters), which is then run once with its constructed weights; empty sources for a model
-                that has trainable parameters is refused here, so a caller cannot silently run random weights.
-        """
-        if not state_sources and any(parameter.numel() for parameter in self._get_model().parameters()):
-            raise PredictorError(
-                "ModelComposite.load() received no checkpoint sources for a model with trainable parameters.",
-                "A weightless model (0 parameters) may run with no checkpoint; a parameterised one may not.",
-                "Pass at least one checkpoint source, or wrap a model that has no parameters.",
-            )
-        self._state_sources = state_sources
-        self._loaded = True
-        self._loaded_state_index = None
-        self._state_cache = {}
-        if len(self._state_sources) == 1:
-            self._ensure_model_loaded(0)
-
-    @torch.inference_mode()
-    def forward(  # type: ignore[override]
-        self,
-        data_dict: dict[tuple[str, bool], torch.Tensor],
-        output_layers: list[str] = [],
-    ) -> list[tuple[str, list[int], torch.Tensor]]:
-        """
-        Perform a forward pass on all model replicas and aggregate their outputs.
-
-        Args:
-            data_dict (dict): A dictionary mapping (group_name, requires_grad) to input tensors.
-            output_layers (list): List of output layer names to extract from each sub-model.
-
-        Returns:
-            list[tuple[str, torch.Tensor]]: Aggregated output for each layer, after applying the reduction.
-        """
-        final_outputs: list[tuple[str, list[int], torch.Tensor]] = []
-        if not self._loaded:
-            raise PredictorError(
-                "ModelComposite.forward() called before load().",
-                "Prediction ran before the composite's checkpoint sources were set.",
-                "Call load(...) first (load([]) for a weightless model).",
-            )
-        # A weightless model (loaded with no checkpoint sources) is a single replica: the model as constructed.
-        n_replicas = len(self._state_sources) or 1
-        if isinstance(self.combine, Mean):
-            sum_acc: dict[str, torch.Tensor] = {}
-            count: dict[str, int] = defaultdict(int)
-            channels: dict[str, list[int]] = defaultdict(list)
-            for model_index in range(n_replicas):
-                for key, tensor in self._model_for_index(model_index)(data_dict, output_layers):
-                    if tensor.dtype == torch.float32:
-                        tensor = tensor.to(torch.float16)
-                    channels[key].append(tensor.shape[1])
-                    if key not in sum_acc:
-                        sum_acc[key] = tensor
-                    else:
-                        sum_acc[key].add_(tensor)
-                    count[key] += 1
-            for key, acc in sum_acc.items():
-                # The sum was folded in place into the first model's output; a lone model's is the
-                # answer as it stands. Dividing by one copied the batch output (56 MiB per
-                # [1, 14, 128^3] fp16 patch, 512 MiB at 122 channels), on every single-model run.
-                final_outputs.append((key, channels[key], acc if count[key] == 1 else acc.div_(count[key])))
-        else:
-            aggregated = defaultdict(list)
-            for model_index in range(n_replicas):
-                for key, tensor in self._model_for_index(model_index)(data_dict, output_layers):
-                    if tensor.dtype == torch.float32:
-                        tensor = tensor.to(torch.float16)
-                    aggregated[key].append(tensor)
-
-            for key, tensors in aggregated.items():
-                # Mean, Median -> [N, C, ...] | Concat -> [N, C*M, ...]
-                final_outputs.append((key, [t.shape[1] for t in tensors], self.combine(tensors)))
-
-        return final_outputs
-
-
-@config()
-class Predictor(DistributedObject):
-    """
-    KonfAI's main prediction controller.
-
-    This class orchestrates the prediction phase by:
-    - Loading model weights from checkpoint(s) or URL(s)
-    - Preparing datasets and output configurations
-    - Managing distributed inference with optional multi-GPU support
-    - Applying transformations and saving predictions
-    - Optionally logging results to TensorBoard
-
-    Attributes:
-        model (Network): The neural network model to use for prediction.
-        dataset (konfai.data.data_manager.DataPrediction): Dataset manager for prediction data.
-        combine_classpath (str): Path to the reduction strategy (e.g., "Mean").
-        autocast (bool): Whether to enable AMP inference.
-        outputs_dataset (dict[str, OutputDataset]): Mapping from layer names to output writers.
-        data_log (list[str] | None): List of tensors to log during inference.
-    """
-
-    def __init__(
-        self,
-        model: ModelLoader = ModelLoader(),
-        dataset: DataPrediction = DataPrediction(),
-        combine: str = "Mean",
-        train_name: str = "name",
-        manual_seed: int | None = None,
-        gpu_checkpoints: list[str] | None = None,
-        autocast: bool = False,
-        channels_last: bool = False,
-        outputs_dataset: dict[str, OutputDatasetLoader] | None = {"default|Default": OutputDatasetLoader()},
-        data_log: list[str] | None = None,
-        check_training_transforms: bool = True,
-    ) -> None:
-        if os.environ["KONFAI_CONFIG_MODE"] != "Done":
-            raise ConfigError("Predictor requires KONFAI_CONFIG_MODE='Done' before initialization.")
-        super().__init__(train_name)
-        self.manual_seed = manual_seed
-        self.dataset = dataset
-        # Auto-patching (VRAM): a per-axis 0 in the user's patch_size marks a FREE axis and opts into
-        # the OOM restart loop: captured before any re-plan materialises concrete sizes over it.
-        patch = dataset.patch
-        self._vram_patch_template: list[int] | None = (
-            [int(size) for size in patch.patch_size]
-            if patch is not None and patch.patch_size is not None and any(size == 0 for size in patch.patch_size)
-            else None
-        )
-        self._vram_patch_candidate: list[int] | None = None
-        # Per-axis input multiple the model needs (its downsampling factor); a free axis is sized to it.
-        self._downsampling_factor: list[int] | None = None
-        module, name = get_module(combine, "konfai.predictor")
-        if module.__name__ == "konfai.predictor":
-            self.combine = getattr(module, name)()
-        else:
-            self.combine = apply_config(f"{konfai_root()}.{combine}")(getattr(module, name))()
-
-        self.autocast = autocast
-        self.channels_last = channels_last
-        self.check_training_transforms = check_training_transforms
-        with startup_clock().phase("model"):
-            self.model = model.get_model(train=False)
-        self.it = 0
-        self.outputs_dataset_loader = outputs_dataset if outputs_dataset else {}
-        self.outputs_dataset = {
-            name.replace(":", "."): value.get_output_dataset(name)
-            for name, value in self.outputs_dataset_loader.items()
-        }
-
-        self.datasets_filename = []
-        self.predict_path = predictions_directory() / self.name
-        per_rank_budget = self.dataset.resolved_budget().per_rank_bytes(node_local_ranks())
-        set_per_rank_budget(per_rank_budget)
-        bound_chunk_cache()
-        for output_dataset in self.outputs_dataset.values():
-            output_dataset.set_memory_budget(per_rank_budget)
-            self.datasets_filename.append(output_dataset.filename)
-            # Rebase under the run directory, re-deriving is_directory: a bare string + "/" would flag an
-            # h5 output as a directory and write the hidden dotfile Predictions/<run>/Dataset/.h5.
-            output_dataset.rebase(self.predict_path)
-        self.data_log = data_log
-        modules = [name for name, _ in self.model.named_modules()]
-        for target in DataLog.parse(self.data_log):
-            if target not in self.dataset.get_groups_dest() and target not in modules:
-                raise PredictorError(
-                    f"Invalid key '{target}' in `data_log`.",
-                    f"This key is neither a destination group from the dataset ({self.dataset.get_groups_dest()})",
-                    f"nor a valid module name in the model ({modules}).",
-                    "Please check your `data_log` configuration,"
-                    " it should reference either a model output or a dataset group.",
-                )
-
-        self.gpu_checkpoints = gpu_checkpoints
-        # Cut the grids with the model's downsampling multiple already known, so each case's free axis
-        # rounds up to a valid input size (the graph (hence the factor) is final before init()).
-        self.dataset.set_free_axis_multiple(self.model.downsampling_factor())
-        self.dataset.prepare()
-        self.model.bind(
-            self.autocast, State.PREDICTION, self.dataset.get_groups_dest(), gpu_checkpoints=self.gpu_checkpoints
-        )
-        # The per-axis multiple a free patch axis rounds up to, read off the model's downsampling graph.
-        self._downsampling_factor = self.model.downsampling_factor()
-        self.output_modules = [name for name, _, _ in self.model.named_module_args_dict()]
-
-        for output_group in self.outputs_dataset.keys():
-            if output_group.replace(";accu;", "") not in self.output_modules:
-                raise PredictorError(
-                    f"The output group '{output_group}' defined in 'outputs_criterions' "
-                    "does not correspond to any module in the model.",
-                    f"Available modules: {self.output_modules}",
-                    "Please check that the name matches exactly a submodule or output of your model architecture.",
-                )
-
-        dataset_groups = {
-            group_src: list(groups_dest.keys()) for group_src, groups_dest in self.dataset.groups_src.items()
-        }
-
-        for name, output_dataset in self.outputs_dataset.items():
-            output_dataset.prepare(name.replace(".", ":"))
-            output_dataset.setup(
-                list(self.dataset.datasets.values()),
-                dataset_groups,
-            )
-
-        if len(self.outputs_dataset) == 0 and not any(
-            network.measure is not None for network in self.model.get_networks().values()
-        ):
-            raise PredictorError(
-                "No prediction outputs or runtime measures are configured.",
-                "Define at least one outputs_dataset entry or enable a network measure.",
-            )
-
-    def setup(self, world_size: int):
-        """
-        Set up the predictor for inference.
-
-        This method performs all necessary initialization steps before running predictions:
-        - Ensures output directories exist, and optionally prompts the user before overwriting existing predictions.
-        - Copies the current configuration file (Prediction.yml) into the output directory for reproducibility.
-        - Dynamically loads pretrained weights from local files or remote URLs.
-        - Wraps the base model into a `ModelComposite` to support ensemble inference.
-        - Initializes the prediction dataloader, with proper distribution across available GPUs.
-
-        Args:
-            world_size (int): Total number of processes or GPUs used for distributed prediction.
-
-        """
-        for dataset_filename in self.datasets_filename:
-            path = self.predict_path / dataset_filename
-            if os.path.exists(path) and len(list(Path(path).rglob("*.yml"))):
-                confirm_overwrite_or_raise(path, "prediction", PredictorError)
-
-            if not os.path.exists(path):
-                os.makedirs(path)
-
-        shutil.copyfile(config_file(), self.predict_path / "Prediction.yml")
-
-        self.model_composite = ModelComposite(self.model, self.combine)
-        if not self.path_to_models and any(parameter.numel() for parameter in self.model.parameters()):
-            # A model WITH weights but no checkpoint would run with random weights and silently produce
-            # garbage: refuse it. A WEIGHTLESS model (0 parameters, e.g. a classical/optimisation engine
-            # such as registration) is legitimate with no checkpoint: it is run once as constructed.
-            raise PredictorError(
-                "No model checkpoint available for prediction.",
-                "This model has trainable weights, so at least one '.pt' checkpoint must be provided (for "
-                "KonfAI Apps, declare it via the 'models' field in app.json).",
-                "Without a checkpoint its weights are random and prediction would silently produce garbage.",
-            )
-        with startup_clock().phase("checkpoint"):
-            self.model_composite.load(self._load())
-        try:
-            self._report_chain_drift()
-        except (OSError, KonfAIError, ValueError, TypeError) as error:
-            # A diagnostic reading someone else's config file never fails the prediction it reports
-            # on; what stopped it is said instead of swallowed.
-            print(f"[KonfAI] the training-chain check did not run: {type(error).__name__}: {error}")
-
-        self.size = len(self.gpu_checkpoints) + 1 if self.gpu_checkpoints else 1
-
-        self.dataloader, _, _ = self.dataset.get_data(world_size // self.size)
-
-    def _report_chain_drift(self) -> None:
-        """Warn when the chain applied to a model input is not the one its checkpoint trained on.
-
-        Same checkpoint, different preprocessing is silent: the run succeeds and only the values are
-        wrong (the Synthesis example shipped ``Standardize(mask: None)`` in training against
-        ``Standardize(mask: MASK)`` here, and paid 409 HU of MAE instead of 98). A legitimate
-        difference exists, so this warns and never refuses, and ``check_training_transforms: false``
-        silences it. Compared against the resolved config the training run left in its ``Statistics``
-        directory, read without writing it back.
-        """
-        if not self.check_training_transforms:
-            return
-        applied = dataset_tree(config_file(), konfai_root())
-        runs: list[tuple[Mapping[str, Any], list[str]]] = []
-        unchecked: list[Path] = []
-        for path_to_model in self.path_to_models:
-            checkpoint = Path(path_to_model)
-            trained = training_dataset_tree(checkpoint) if checkpoint.is_file() else None
-            if trained is None:
-                unchecked.append(checkpoint)
-                continue
-            # Folds of one experiment spell the same chains: compare each distinct one once.
-            for known, names in runs:
-                if known == trained:
-                    names.append(checkpoint.parent.name)
-                    break
-            else:
-                runs.append((trained, [checkpoint.parent.name]))
-        if unchecked:
-            print(
-                f"[KonfAI] the training-chain check did not run for {len(unchecked)} checkpoint(s):"
-                f" no resolved training config beside '{unchecked[0]}'"
-                " (a TRAIN run keeps one in Statistics/<train_name>/)."
-            )
-        for trained, names in runs:
-            differences = input_chain_differences(trained, applied)
-            if not differences:
-                continue
-            print(f"[KonfAI] WARNING: this run preprocesses a model input differently from {', '.join(names)}:")
-            for difference in differences:
-                print(f"[KonfAI]   {difference}")
-            print(
-                "[KonfAI] The same checkpoint on differently preprocessed inputs predicts wrong values"
-                " without failing. Set 'check_training_transforms: false' under Predictor once the"
-                " difference is deliberate."
-            )
-
-    def set_models(self, path_to_models: list[Path | str]) -> None:
-        self.path_to_models = path_to_models
-
-    def _load(self) -> list[dict[str, Any] | Path | str]:
-        """
-        Resolve checkpoint sources for ensemble prediction.
-
-        This method handles both remote and local model sources:
-        - If the model path is a URL (starting with "https://"), it eagerly downloads and loads the state dict
-          once because re-fetching it every batch would be prohibitively slow.
-        - If the model path is local:
-            - it keeps only the checkpoint path and lets `ModelComposite` stream weights into a single model
-              instance during prediction to reduce memory pressure.
-
-        Returns:
-            list[dict[str, dict[str, torch.Tensor]] | Path | str]: A list of checkpoint sources, one per model.
-
-        Raises:
-            Exception: If a model path does not exist or cannot be loaded.
-        """
-        state_dicts = []
-        for path_to_model in self.path_to_models:
-            if isinstance(path_to_model, str) and path_to_model.startswith("https://"):
-                try:
-                    state_dicts.append(
-                        torch.hub.load_state_dict_from_url(url=path_to_model, map_location="cpu", check_hash=True)
-                    )
-                except Exception as exc:
-                    raise Exception(f"Model : {path_to_model} does not exist !") from exc
-            elif Path(path_to_model).exists():
-                state_dicts.append(Path(path_to_model))
-            else:
-                raise ValueError(f"Invalid model path entry: {path_to_model}")
-        return state_dicts
-
-    def run_process(
-        self,
-        world_size: int,
-        global_rank: int,
-        local_rank: int,
-        dataloaders: list[DataLoader],
-    ):
-        """
-        Launch prediction on the given process rank.
-
-        Args:
-            world_size (int): Number of model replicas sharding the data: the spawned process count
-                already divided by the model-parallel size (``gpu_checkpoints``), NOT the GPU count.
-            global_rank (int): Rank of the current process.
-            local_rank (int): Local device rank.
-            dataloaders (list[DataLoader]): List of data loaders for prediction.
-        """
-
-        model_composite = (
-            Network.to(self.model_composite, local_rank * self.size)
-            if len(cuda_visible_devices())
-            else self.model_composite
-        )
-        if self.channels_last:
-            Network.set_channels_last(model_composite)
-        if len(cuda_visible_devices()):
-            # Co-locate the output writers with the model so their reduction/transforms know the GPU.
-            for output_dataset in self.outputs_dataset.values():
-                output_dataset.to(local_rank * self.size)
-        model_composite = Model(model_composite)
-        device = local_rank * self.size if len(cuda_visible_devices()) else None
-        dataloader = dataloaders[0]
-        # Round a free patch axis up to the model's valid input multiple before the first attempt, so
-        # the network's encoder/decoder skips align instead of crashing on a non-divisible extent (the
-        # border padding fills the round-up, cropped back after the forward). A whole-axis extent still
-        # too large for VRAM OOMs into the shrink loop below, which keeps the size valid too.
-        if self._vram_patch_candidate is None:
-            sized = size_free_axes(
-                self._vram_patch_template, self.dataset.worst_case_shape(), self._downsampling_factor
-            )
-            if sized is not None:
-                self._vram_patch_candidate = sized
-                self.dataset.replan_patch(sized)
-                dataloader = self.dataset.get_data(world_size)[0][global_rank][0]
-        while True:
-            try:
-                with _Predictor(
-                    world_size,
-                    global_rank,
-                    local_rank,
-                    self.autocast,
-                    self.predict_path,
-                    self.data_log,
-                    self.outputs_dataset,
-                    model_composite,
-                    dataloader,
-                ) as p:
-                    p.run()
-                return
-            except torch.cuda.OutOfMemoryError:
-                # The restart loop IS the sizing iteration (no probe phase): the run that just OOMed
-                # already measured the step's transient for free. Read it BEFORE the reset (the peak
-                # still includes the resident accumulators on both sides of the difference), free the
-                # in-flight state: open streamed sinks abort and remove their partial entries, so a
-                # reader never sees a half-written volume even when the OOM is fatal, then read the
-                # honest free VRAM.
-                measured = vram.transient_at_oom(device)
-                for output_dataset in self.outputs_dataset.values():
-                    output_dataset.reset()
-                if self._vram_patch_template is None:
-                    raise  # no free axis declared: not auto-patched
-                candidate = self._shrunken_patch(measured, vram.usable_after_oom(device))
-                if candidate is None:
-                    raise
-                vram.reset_peak(device)
-                print(
-                    f"[KonfAI] VRAM: rank {global_rank} ran out of memory -> "
-                    f"re-planning the free patch axes to {candidate} and restarting this rank's cases."
-                )
-                self._vram_patch_candidate = candidate
-                self.dataset.replan_patch(candidate)
-                dataloader = self.dataset.get_data(world_size)[0][global_rank][0]
-
-    def _shrunken_patch(self, measured: int | None, usable: float) -> list[int] | None:
-        """One shrink step for the free patch axes after a CUDA OOM (``None`` = not auto, or floor).
-
-        The first OOM starts from the worst prepared case at full extent (the size the failed grid
-        effectively ran); later ones shrink the current candidate further. When the framework picks
-        the size, it must also leave the blend on the GPU: the accumulation footprint is RESERVED
-        beside the forward, so the sized patch passes the accumulation gate. Only when that reserve
-        fits at no size (or cannot be priced) is the forward sized alone: the gate's memory-safe
-        CPU blend absorbs that case.
-        """
-        if self._vram_patch_template is None:
-            return None
-        worst = self.dataset.worst_case_shape()
-        if worst is None:
-            return None
-        candidate = self._vram_patch_candidate or concretize_patch_size(
-            self._vram_patch_template, worst, self._downsampling_factor
-        )
-        reserve = self._accumulation_reserve(candidate, worst)
-        snap = self._downsampling_factor
-        if reserve is not None:
-            shrunk = vram.next_patch_candidate(
-                candidate, self._vram_patch_template, worst, measured, usable - reserve, snap
-            )
-            if shrunk is not None:
-                return shrunk
-        return vram.next_patch_candidate(candidate, self._vram_patch_template, worst, measured, usable, snap)
-
-    def _accumulation_reserve(self, candidate: list[int], worst: list[int]) -> float | None:
-        """Bytes each case keeps resident while its patches accumulate, per output writer: the
-        streamed window (one patch extent x the cross-section) when the writer will stream --
-        single augmentation, voxel-local reduction: the assembled volume otherwise. ``None``
-        when a writer's channels cannot be read off the model trace (no reserve, gate decides).
-        """
-        trace = {name: args.out_channels for name, _, args in self.model.named_module_args_dict()}
-        elem = 2  # ModelComposite casts float32 outputs to float16 before accumulation
-        reserve = 0.0
-        for name, writer in self.outputs_dataset.items():
-            out_channels = trace.get(name.replace(";accu;", ""))
-            if not out_channels:
-                return None
-            if isinstance(self.combine, Concat):
-                out_channels *= max(1, len(self.path_to_models))
-            nb_augmentation = max(1, writer.nb_data_augmentation)
-            streams = nb_augmentation == 1 and writer.reduction.voxel_local
-            voxels = candidate[0] * np.prod(worst[1:], dtype=np.int64) if streams else np.prod(worst, dtype=np.int64)
-            reserve += float((out_channels + 1) * voxels * elem * nb_augmentation)
-        return reserve
-
-    def __str__(self) -> str:
-        params = {
-            "model": self.model,
-            "dataset": self.dataset,
-            "combine": self.combine,
-            "train_name": self.name,
-            "manual_seed": self.manual_seed,
-            "gpu_checkpoints": self.gpu_checkpoints,
-            "autocast": self.autocast,
-            "outputs_dataset": self.outputs_dataset,
-            "data_log": self.data_log,
-        }
-        return str(params)
-
-    def __repr__(self) -> str:
-        return str(self)
-
-
-def build_predict(
-    models: list[Path],
-    prediction_file: Path | str | dict = Path("./Prediction.yml").resolve(),
-    predictions_dir: Path | str = Path("./Predictions").resolve(),
-) -> DistributedObject:
-    """
-    Build and return the configured prediction workflow without executing it.
-
-    Parameters
-    ----------
-    models : list[Path]
-        One or more checkpoint files to load for prediction.
-    prediction_file : Path | str, optional
-        Prediction configuration file.
-    predictions_dir : Path | str, optional
-        Directory where prediction outputs are written.
-
-    Returns
-    -------
-    DistributedObject
-        Configured predictor object ready to be executed by the runtime wrapper.
-    """
-    configure_workflow_environment(
-        config_path=prediction_file,
-        root="Predictor",
-        state=State.PREDICTION,
-        path_env={"KONFAI_PREDICTIONS_DIRECTORY": predictions_dir},
-    )
-    os.environ["KONFAI_CONFIG_MODE"] = "Done"
-    with strict_config("Predictor", refuse=False):
-        predictor = apply_config()(Predictor)()
-    predictor.set_models(models)
-    return predictor
-
-
-@run_distributed_app
-def predict(
-    models: list[Path],
-    overwrite: bool = False,
-    gpu: list[int] | None = cuda_visible_devices(),
-    cpu: int = 1,
-    quiet: bool = False,
-    tensorboard: bool = False,
-    prediction_file: Path | str | dict = Path("./Prediction.yml").resolve(),
-    predictions_dir: Path | str = Path("./Predictions").resolve(),
-) -> DistributedObject:
-    """
-    Build and execute the configured prediction workflow.
-
-    This compatibility wrapper preserves the historical CLI-facing API while
-    delegating the pure build step to :func:`build_predict`.
-    """
-    del overwrite, gpu, cpu, quiet, tensorboard
-    return build_predict(
-        models=models,
-        prediction_file=prediction_file,
-        predictions_dir=predictions_dir,
-    )
